Create VM,BM,I
Созданы ViewModel
Созданы BindingModel
Созданы Интерфейсы

+++ ITShopBusinessLogic/ITShopBusinessLogic/BindingModels/ProductBindingModel.cs
@@ -7,10 +7,10 @@ namespace ITShopBusinessLogic.BindingModels
     public class ProductBindingModel
     {
         public int? Id { get; set; }
-      //public int SpecificationsId {get;set;}
+        //public int SpecificationsId {get;set;}
         public string Name { get; set; }
         public DateTime ReleaseYear { get; set; }
         public DateTime WarrantyEnd { get; set; }
-        public double ProductPrice { get; set; }
+        public decimal ProductPrice { get; set; }
     }
 }
